autoapply/backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Auth
PyJWT==2.10.1
httpx==0.28.1
clerk-backend-api==1.4.0

# AI
anthropic==0.40.0

# Resume parsing
PyMuPDF==1.25.1
python-docx==1.1.2

# Job scraping
jobspy==1.1.80

# Browser automation
playwright==1.49.1
playwright-stealth==1.0.6

# Queue
celery[redis]==5.4.0
redis==5.2.1
flower==2.0.1

# PDF generation
weasyprint==63.1

# Utilities
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
Pillow==11.0.0
pandas==2.2.3
loguru==0.7.3
tenacity==9.0.0
